Fixing DB API Response (WORKING)
