+++ server/package.json
@@ -22,7 +22,6 @@
   "dependencies": {
     "express": "^4.17.3",
     "mongoose": "^6.3.0",
-    "mysql2-promise": "^0.1.4",
     "mysqli": "^3.1.4"
   }
 }
